Refactor following parallel sampling change (#121)
* unify test_engine and test_parallel_sampling

* moved cache manager to its own file

* clean up

* fix
